Extract `getSelectedItems`
