+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbar.tsx
@@ -6,8 +6,6 @@ import { LocationDescriptor } from "history";
 import * as Urls from "metabase/lib/urls";
 import { closeNavbar, openNavbar } from "metabase/redux/app";
 
-import { coerceCollectionId } from "metabase/collections/utils";
-
 import { getQuestion } from "metabase/query_builder/selectors";
 import { getDashboard } from "metabase/dashboard/selectors";
 
@@ -24,6 +22,7 @@ import {
   MainNavbarDispatchProps,
   SelectedItem,
 } from "./types";
+import getSelectedItems from "./getSelectedItems";
 import { NavRoot, Sidebar } from "./MainNavbar.styled";
 
 interface StateProps {
@@ -78,67 +77,16 @@ function MainNavbar({
     };
   }, [isOpen, openNavbar, closeNavbar]);
 
-  const selectedItems = useMemo<SelectedItem[]>(() => {
-    const { pathname } = location;
-    const { slug } = params;
-    const isCollectionPath = pathname.startsWith("/collection");
-    const isUsersCollectionPath = pathname.startsWith("/collection/users");
-    const isQuestionPath = pathname.startsWith("/question");
-    const isModelPath = pathname.startsWith("/model");
-    const isDataAppPath = Urls.isDataAppPath(pathname);
-    const isDataAppPagePath = Urls.isDataAppPagePath(pathname);
-    const isDashboardPath = pathname.startsWith("/dashboard");
-
-    if (isCollectionPath) {
-      return [
-        {
-          id: isUsersCollectionPath ? "users" : Urls.extractCollectionId(slug),
-          type: "collection",
-        },
-      ];
-    }
-    if (isDataAppPagePath) {
-      return [
-        {
-          id: parseInt(params.pageId as string),
-          type: "data-app-page",
-        },
-      ];
-    }
-    if (isDataAppPath) {
-      return [
-        {
-          id: Urls.extractEntityId(slug),
-          type: "data-app",
-        },
-      ];
-    }
-    if (isDashboardPath && dashboard) {
-      return [
-        {
-          id: dashboard.id,
-          type: "dashboard",
-        },
-        {
-          id: coerceCollectionId(dashboard.collection_id),
-          type: "collection",
-        },
-      ];
-    }
-    if ((isQuestionPath || isModelPath) && question) {
-      return [
-        {
-          id: question.id(),
-          type: "card",
-        },
-        {
-          id: coerceCollectionId(question.collectionId()),
-          type: "collection",
-        },
-      ];
-    }
-    return [{ url: pathname, type: "non-entity" }];
-  }, [location, params, question, dashboard]);
+  const selectedItems = useMemo<SelectedItem[]>(
+    () =>
+      getSelectedItems({
+        pathname: location.pathname,
+        params,
+        question,
+        dashboard,
+      }),
+    [location, params, question, dashboard],
+  );
 
   return (
     <Sidebar className="Nav" isOpen={isOpen} aria-hidden={!isOpen}>
